INC-735: Prison API doesn't swallow exceptions
Before this change the methods in the `PrisonApiService` that returned
return a `Flow` were handling the upstream `404 Not Found` by returning
an empty flow to the caller.

Although this sounds sensible (we try to return a Flow)
the problem with this approach is that the caller may get a `Flow` instance
and think everything is fine, then call `first()` on this empty `Flow`
and trigger another exception.

The Incentives API client gets a `500 Internal Server Error` and we lost
details about the original source of the problem.

After this change **all** exceptions will bubble up (non-`NotFound`
exceptions were already re-thrown before so this shouldn't make a difference)

`NotFound` exceptions are not handled (see previous commit)