Removed unnecessary `let {}`/`with()`
When only calling a method use these is overkill and makes the code look
more complicated than it is.

e.g.

```kotlin
prisonerIepLevelRepository.findById(id)?.let {
  with(it) {
    translate()
  }
}
```

this is just calling translate on the result of `findById()`, it reads
much harder than `prisonerIepLevelRepository.findById(id)?.translate()`
but the two are doing the same thing.