INC-737: Fixed last tests related to `sequence` being removed
API response doesn't include the `sequence` values anymore (not even
when passing the param to get data from NOMIS/Prison API)

NOTE: Removing `sequence` from the API responses is a **BREAKING CHANGE**
We need to check other services/consumers of the API to determine whether
`sequence` is used (preliminary/cursory look seems to suggest it is not
used).
We also need to give other teams an heads up and double check if they're
using the `sequence` field in the response or even are planning to use it
in the future.