INC-737: Remove `sequence` from model (BREAKING change) (#161)
* Removed unnecessary `toList()` on `filter {}` result

filtering already return a `List<T>`, no reason to "convert" to a list
again.

* INC-737: DB migration to drop 'sequence' column in prisoner_iep_level table

This will also drop the associated index and unique constraint on
(bookingId, sequence).

* INC-737: Update PrisonerIepLevel repository to work without sequence

The `sequence` column is gone. There were a couple of methods that
were referring it in the name, and also some tests passing this
`sequence` value when saving test data.

These were updated.

TODO: This is still WIP and other part of the codebase still need to be
updated.

* INC-737: Removed logic to increase PrisonerIepLevel sequence

From `PrisonerIepLevelReviewService`. The natual order is determined by
the IEP review time.

* INC-737: Fixed tests using `sequence`

Application now compiles again. Some tests may still fail and there may
be some code that needs updating.

* INC-737: Fixed test using `sequence` after rebase

* INC-737: Fixed last tests related to `sequence` being removed

API response doesn't include the `sequence` values anymore (not even
when passing the param to get data from NOMIS/Prison API)

NOTE: Removing `sequence` from the API responses is a **BREAKING CHANGE**
We need to check other services/consumers of the API to determine whether
`sequence` is used (preliminary/cursory look seems to suggest it is not
used).
We also need to give other teams an heads up and double check if they're
using the `sequence` field in the response or even are planning to use it
in the future.

* INC-737: Renamed migration file