INC-738: Refactoring to rename IepMigration to SyncPostRequest
The migration endpoint request format is the same of the sync POST endpoint
request. It makes sense to use the same data class.

Also, conceptually the migration endpoint is there to syncronise data
from NOMIS similarly to the sync POST endpoint.

For this reasons I've renamed the `IepMigration` class to `SyncPostRequest`
and renamed similar variables and utility methods accordingly.