INC-739: Made POST sync endpoint test less brittle
Once we add more tests/records for the PATCH sync endpoint this test
would have failed because it was assuming only one record was in the DB
at the time of creation. It also didn't check the response had an `id`
field (something the sync consumers will likely be interested in).

I've updated the test to:
1. parse the response to read the ID (hence verifying this is present)
2. use the `GET /iep/reviews/id/$id` endpoint to verify this record
   was persisted as expected