INC-739: PATCH /sync/booking/{bookingId}/id/{id} endpoint (#174)
* INC-739: Added `SyncPatchRequest` data type

As per ticket only `iepTime`, `comment` and `current` can be updated
and these fields are all optional (e.g. you could patch only one of these
if you needed to).

* INC-739: Made POST sync endpoint test less brittle

Once we add more tests/records for the PATCH sync endpoint this test
would have failed because it was assuming only one record was in the DB
at the time of creation. It also didn't check the response had an `id`
field (something the sync consumers will likely be interested in).

I've updated the test to:
1. parse the response to read the ID (hence verifying this is present)
2. use the `GET /iep/reviews/id/$id` endpoint to verify this record
   was persisted as expected

* INC-739: Added tests for sync PATCH endpoint

TODO: More tests to the service needed (e.g. to test domain event and
audit event)

* INC-739: Added "updated" types to event and audit message

- Added `IEP_REVIEW_UPDATED` to `AuditType`
- Added `IncentivesDomainEventType` enum with `IEP_REVIEW_INSERTED`/
  `IEP_REVIEW_UPDATED`
  - This enum has String values for the corresponding domain event types
    constants
- Be explicit about what domain event type and audit type are sent

* INC-739: Added `PATCH /iep/sync/booking/{bookingId}/id/{id}` endpoint

- this is to update existing IEP reviews by ID.
- as per ticket, only `comment`, `iepTime` and `current` fields can be change.
- at least one of these 3 fields must be provided
- `incentives.iep-review.updated` domain event is published
- `IEP_REVIEW_UPDATED` audit event is published

* INC-739: Fix swagger docs

- specify `RequestBody`'s `@Schema(implementation=)` so that example request
  body is populated and useful
- consistency use same values for same fields across all the swagger example
  so that request/response mostly match (Swagger seems to ignore DateTime
  example for some reason)
- have the `example` param as last in `@Schema` for consistency

* INC-739: tests for service's `handleSyncPatchIepReviewRequest()`

Useful to test correct domain event and audit event are published

* INC-739: Fix import/spaces linting problems

* INC-739: Fixed flaky tests caused by small `iepTime` differences

Hopefully explicitly use `.format(DateTimeFormatter.ISO_DATE_TIME)` instead
of just interpolating the datetime should fix the flaky tests we're seeing
in CI and locally sometimes, e.g.

```
java.lang.AssertionError: iepTime
Expected: 2022-09-19T10:13:13.747634283
     got: 2022-09-19T10:13:13.747634
```

PS: Also removed leftover `// TODO` comment

* INC-739: Attempt to fix flaky tests related to datetime formatting

The previous attempt did not work but we can be even more explicit and
specify the exact format using `DateTimeFormatter.ofPattern()` which should
solve this problem.

* INC-739: Attempt to fix flaxy tests

Removed `"iepTime"` from response check as the way the datetime is
serialised locally and on CircleCI somehow seems to be different and it
sometimes fails because of small fractions of a second differences.

**NOTE**: This is not ideal (we may want to revisit this) but at least
there are also tests for the service which should cover the presense of
the `iepTime` field in the response body (as service method return value
gets serialised straight as response)