INC-760: Separate publishing of domain event from audit events
These two kind of events have been always published together when an IEP
review is created/updated.
This may not be always the case as we add a feature flag to conditionally
publish a domain event OR make a `POST` request to Prison API (in the 2
add IEP endpoints).

Regardless of how we propagate this change to NOMIS (via domain event sync
or request to Prison API) we still want to publish the audit event.

So it makes sense to separate the publication of these 2 events.