INC-815: Basic next review date calculation (#187)
* Persist upstream updates to trivy exclusions

* Indent read-me markdown correctly

* Pass `stubIEPSummaryForBooking` stub’s `bookingId` parameter out in response like other stubs do

* Refactor `daysSinceReview` code
At present, `daysSinceReview` can come from 2 sources: directly from prison api OR calculated from incentives db values.

The `IepSummary` data class required that `daysSinceReview` be passed in on construction and yet had a method to calculate the same value from history also provided in the constructor as `iepDetails`. However, `iepDetails` is empty when loaded without details – this would lead to the field and method disagreeing, with the method throwing an exception.

Regardless, `daysSinceReview` only requires the latest review date which is always present irrespective of data source or whether full details were loaded. Therefore, it can be made into a derived property which is always calculated on access (i.e. during serialisation to JSON).

* Fix incorrect test data
`iepDate` & `iepTime` of `IepSummary` are always the same as the first child inside the nested `iepDetails`

* Fix prison api stubs: `iepDate` should not have time component

* Calculate simplest form of next review date
…which applies to the majority of the population: those not on basic, nor newly in prison who have not been transferred since the last review

* Use clearer variable name inside `daysSinceReview` and `daysOnLevel`