INC-827: Improve `daysOnLevel` calculation (#190)
* Improve `daysOnLevel` calculation
Previously, it silently returned 0 when history was unavailable (i.e. `IepSummary` was loaded "without details") instead of indicating that there was a problem.
It was only used in one place (as an intermediate object) and was not serialised as part of the `IepSummary` schema.

Throwing an exception seems prudent to indicate that there is an _implementation_ problem (wrong details passed in).
The calculation has also been moved closer to where it's used. It's less error prone than being a method that looked reusable.

* Make all of `IncentiveSummaryService`'s helper methods private so they don't look like they're part of the external api