INC-760: `reviewAddedSyncMechanism` feature flag
Added `feature.reviewAddedSyncMechanism` feature flag which determine
how a new IEP review is propagated to other systems.

This can be `PRISON_API` (default value currently used in `preprod`/`prod`)
meaning an HTTP request to Prison API will be made to update NOMIS.

Alternatively this value can be `DOMAIN_EVENT` (value used in `dev`)
meaning that instead of updatind NOMIS via Prison API a domain event is
published and the subscriber will be responsible for updating NOMIS (or
doing whatever they need to do with this piece of information).

In `dev` the domain event will be used as there is a sync service listing
to these domain events (the fact we're currently also updating NOMIS by
making a request to Prison API is causing duplcated records)

So only one of the other will happen. Long term, when this sync service
will be rolled out to `preprod`/`prod` we can just publish the domain
event for all environment and possibly remove this feature flag.

NOTE: The corresponding environment variable is `FEATURE_REVIEWADDEDSYNCMECHANISM`
and the allowed values are `DOMAIN_EVENT` and `PRISON_API` (case sensitive
as they're mapped into the corresponding enum value)