INC-760: Feature flag to either publish domain event or call Prison API on new IEP review (#197)
* INC-760: Moved events logic into service method

These two add IEP review endpoints share most of their logic, including
the publishing of domain event and audit event:
- `POST /iep/reviews/prisoner/{prisonerNumber}`
- `POST /iep/reviews/booking/{bookingId}`

Some of this logic was in the resource class, this is now moved into the
service class and the duplication is also removed.

* INC-760: Separate publishing of domain event from audit events

These two kind of events have been always published together when an IEP
review is created/updated.
This may not be always the case as we add a feature flag to conditionally
publish a domain event OR make a `POST` request to Prison API (in the 2
add IEP endpoints).

Regardless of how we propagate this change to NOMIS (via domain event sync
or request to Prison API) we still want to publish the audit event.

So it makes sense to separate the publication of these 2 events.

* INC-760: Refactoring: Better name for `addIepLevel()`

This method adds an IEP review given a `PrisonerAtLocation` parameter,
so renamed to `addIepReviewForPrisonerAtLocation()`.

NOTE: This is the method that both `addIepReview()` overloaded methods
invoke.

* INC-760: Add tests for `addIepReview()` methods

Their logic wasn't tested and it's going to change once we add the flag
to EITHER publish a domain event OR make a POST request to Prison API.

So it's good to test what's already there.

* INC-760: `reviewAddedSyncMechanism` feature flag

Added `feature.reviewAddedSyncMechanism` feature flag which determine
how a new IEP review is propagated to other systems.

This can be `PRISON_API` (default value currently used in `preprod`/`prod`)
meaning an HTTP request to Prison API will be made to update NOMIS.

Alternatively this value can be `DOMAIN_EVENT` (value used in `dev`)
meaning that instead of updatind NOMIS via Prison API a domain event is
published and the subscriber will be responsible for updating NOMIS (or
doing whatever they need to do with this piece of information).

In `dev` the domain event will be used as there is a sync service listing
to these domain events (the fact we're currently also updating NOMIS by
making a request to Prison API is causing duplcated records)

So only one of the other will happen. Long term, when this sync service
will be rolled out to `preprod`/`prod` we can just publish the domain
event for all environment and possibly remove this feature flag.

NOTE: The corresponding environment variable is `FEATURE_REVIEWADDEDSYNCMECHANISM`
and the allowed values are `DOMAIN_EVENT` and `PRISON_API` (case sensitive
as they're mapped into the corresponding enum value)

* Fixed rebase problems caused by refactorings in `main`

- `persistPrisonerIepLevel()` was renamed to `persistSyncPostRequest()`
- `SnsService#sendIepReviewEvent()` now gets an additional argument for
  the prisoner number (2nd argument)

* Use lowercase kabab format for feature flag property name

To a more readable `feature.review-added-sync-mechanism` as suggested in
the following Spring Boot document and consistent to many of the other properties:

https://github.com/spring-projects/spring-boot/wiki/Relaxed-Binding-2.0

NOTE: The corresponding environment variable doesn't change and is still
`FEATURE_REVIEWADDEDSYNCMECHANISM` (as dots are replaces with spaces but
anything else is dropped)

* Linting fixes

* INC-760: Removed duplication between `addIepReview()`s' tests

There are 2 `addIepReview()` methods, one by `bookingId`, the other by
`offenderNumber`. They're mostly the same (they call the same underlying
`addIepReviewForPrisonerAtLocation()` method).

Their tests are very similar but aside from the slightly different setup
the tests for the common functionality are not duplicated now.