INC-823: enforce prisoner_iep_level contains one record of current=true per booking ID
See https://stackoverflow.com/a/34495621 for why folks suggest to use an index rather than constraint. Adding this because we may encounter a race condition where we attempt to create multiple IEP Levels with current=true, as our service has multiple entry points:
- create through DPS
- sync endpoint
- processing domain event message
Which means we be creating records from different transactions. Enforcing this logic on the table itself should ensure in that scenario then repository.save will fail with a DataIntegrityViolationException, and the transaction can be reattempted at a later time