INC-872: sync DELETE endpoint (#207)
* INC-872: sync DELETE endpoint resource tests

* INC-872: Added service tests for sync DELETE endpoint

* INC-872: Added sync DELETE endpoint

`DELETE /iep/sync/booking/{bookingId}/id/{id}` endpoint

* INC-872: Add domain/audit `IEP_REVIEW_DELETED` event types

* INC-872: Basic handling of sync DELETE endpoint requests

- ✅ checks bookingId is valid
- ✅ deletes record
- ✅ publishes domain/audit events
- ❌ **NOTE**: Does **not** set another IEP review as current if the deleted
  record had `current = true`

The deleted record may be the one with `current = true`, in that case we
should probably set another IEP review to current, e.g. the latest IEP
review.

* INC-872: sync DELETE endpoint logic for current flag

When an IEP review is deleted, if this had the current flag set to true
this ensures there will be another record (if any) with current = true.

The then latest IEP review for that bookingId will become the current IEP
review.