INC-818: next review date takes 'Basic' into consideration
Requirements/rules
==================

> PF 5.16 Prisoners placed on Basic must be reviewed within 7 days and if
> not suitable to return to Standard level further reviews must be
> undertaken at least every 28 days thereafter

Ticket scope
============

There are other rules but this ticket is about improving the next review
date for prisoners on 'Basic' level.

Technical implementation
========================

The `nextReviewDate` property in `IepSummary` has now a getter which
delegates the calculation of the next review date to the `NextReviewDateService`.

The idea is that this service is stateless and given all the information
it needs, determines the next review date.

**NOTE**: At the moment this information is the last review date/level
and the reviews history (`iepDetails`) but in the future it may need more
information, e.g. ACCT or other.