INC-818: Fix for "no iepDetails" edge case
It was theoretically possible for one endpoint to pass withDetails=false
causing the `nextReviewDate` for someone with the last two reviews levels
at 'Basic' to return a next review date of +7 days instead of +28 days.

All places where an `IepSummary` is instanciated have been updated to
1. always get the iepDetails/reviews history
2. set the nextReviewDate
3. when the method has a withDetails parameter and this was passed as false
   `iepDetails` are replaced with an empty list so that the API response
   doesn't contain it.

Also, the `nextReviewDate` in the `IepSummary` class is now a property
instead of a getter.