INC-818: next review date takes 'Basic' into consideration (#224)
* Renamed `incentivesDataSourceOfTruth` feature flag for consistency

The Spring Boot property was still called `feature.incentives-data-source-of-truth`
but the corresponding method was called something else.

* INC-818: next review date takes 'Basic' into consideration

Requirements/rules
==================

> PF 5.16 Prisoners placed on Basic must be reviewed within 7 days and if
> not suitable to return to Standard level further reviews must be
> undertaken at least every 28 days thereafter

Ticket scope
============

There are other rules but this ticket is about improving the next review
date for prisoners on 'Basic' level.

Technical implementation
========================

The `nextReviewDate` property in `IepSummary` has now a getter which
delegates the calculation of the next review date to the `NextReviewDateService`.

The idea is that this service is stateless and given all the information
it needs, determines the next review date.

**NOTE**: At the moment this information is the last review date/level
and the reviews history (`iepDetails`) but in the future it may need more
information, e.g. ACCT or other.

* INC-818: PR amendment: make "Basic" a constant

* INC-818: PR amendments, correct tests typos

* INC-818: PR amendments: Use correct Exception class

* INC-818: Fix for "no iepDetails" edge case

It was theoretically possible for one endpoint to pass withDetails=false
causing the `nextReviewDate` for someone with the last two reviews levels
at 'Basic' to return a next review date of +7 days instead of +28 days.

All places where an `IepSummary` is instanciated have been updated to
1. always get the iepDetails/reviews history
2. set the nextReviewDate
3. when the method has a withDetails parameter and this was passed as false
   `iepDetails` are replaced with an empty list so that the API response
   doesn't contain it.

Also, the `nextReviewDate` in the `IepSummary` class is now a property
instead of a getter.