INC-816: Wire up offender search service to return a single offender by prisoner number (#227)
* Give offender search response a better name as there are other responses from other endpoints
* Wire up offender search service to return a single offender by prisoner number which will be needed for future "date of next review" calculations.
* Use system token to access offender search
Essentially, we only require the `MAINTAIN_IEP` role, and this makes sense for the data we actually return to the caller. But to calculate next review dates, we need to load more information on the given offender (a behind-the-scenes implementation detail) and the calling user may not have the roles to do so.