INC-852: next review date calculation uses open ACCT only for Basic
Discussed with Steve (BA) and he confirmed that the ACCT/Young People
rules only apply for someone on Basic, not for other levels.

I've updated the test and the calculation to reflect this.

PS: I've also refactored the class to get the input as a constructor
parameter which means I can now have nicely named methods for concepts
like "is on basic" or "is new prisoner" without having to destructure
the input data class in various places.