INC-820: Next review date calculation only takes in consideration real reviews (#245)
* INC-820: Moved "is real review" logic into an interface

We have a couple of classes that have a `reviewType` field and for instances
of these we want to be able to determine whether a "review" (whether it is
an `PrisonerIepLevel` or `IepDetail`, etc...) is real or one of those
system-generated "reviews" (e.g. admissions, transfers, etc...)

I could have put this logic in the `ReviewType` enum but this doesn't seem
right as whether a review is a real one or not is a characteristic of the
review not the `reviewType` field, hence the use of an interface.

I've also moved the `ReviewType` class from under `jpa` into the `dto`
package as this is not a concept strictly related to persistance and now
that this logic is used in non-jpa classes it makes sense to move it.

Notes on the logic
------------------
Reviews migrated from NOMIS would have a type of `MIGRATED`, this means
we can't really tell whether these are "real" reviews. However decision
has been made that these are considered to be "real", e.g. they'll not be
ignored by various bits of code (this logic was already present before,
I just moved it into the `IsRealReview` interface).

Similarly, data could come from NOMIS has no "review type", as this is
a concept we introduced in our service (`IepDetail` have an optional
`reviewType` field, when data comes from NOMIS, which is the case in
production at the moment, this would be `null`)

In practice this means that if data is coming from NOMIS (as it's the case
in production at the moment) we cannot properly discriminate between
real reviews and other types of entries.
Similarly to the case above about `MIGRATED`, if data is coming from NOMIS
we don't ignore it and we assume these reviews are "real".

* INC-820: Next review date only takes in consideration real reviews

If reviews come from our database they'll have a `reviewType`.
- if this is `REVIEW` they're real reviews
- if this is `MIGRATED` these were migrated from NOMIS and we assume they're
  "real"
- when the data is coming from Prison API/NOMIS (as it currently does
for the time being in production) we don't have the review type information
but we consider them real reviews as in the case of `reviewType='MIGRATED'`
(ultimately this will not matter as we'll only get data from our DB and
these "legacy" reviews will have 'MIGRATED' type)

* INC-820: Lint fixes