INC-817: next_review_date DB table/jpa class and repository
Added a table to store the `nextReviewDate` by `bookingId`.

`bookingId` is the natural primary key (PK) and, unsurprisingly, the other
main columnis the `nextReviewDate` date.

For each `bookingId` there would be only record max, this is a constraint
enforced by the DB because of the fact this is the PK.

We'd likely retrieve "next review dates" for a list of bookingIds.
This is provided by `CoroutineCrudRepository`'s `findAllById()` so we don't
have to add this to the repository when that time comes (I've added a test
for this even tho it's probably overkill as I'd imagine the API for
`CoroutineCrudRepository` is very stable).