INC-817: Class to update next review dates in the database
This will be used to recalculate/update next review dates in the database.

This is necessary when:
1. trying to get a next review date for some bookingIds but some of them
   don't have this value yet (so we need to calculate it)
2. something changes and next review date needs to be updated, e.g.
   a new review is added or someone ACCT status changed

This method gets a list of Offender Search prisoners as this information
is needed to calculate the next review date.

NOTE: At the moment only an `updateMany()` method is available but we can
add another method to `update()` a single offender later on which is
probably what we need in most cases on admission/transfer or when a new
review is created.