INC-817: Class to get next review dates
This class tries to get these from the database first, for bookingIds
without a `NextReviewDate` DB record it calculate/persist the next review
dates.

After a call to this class all the offenders provided will have a
`NextReviewDate` DB record.

All the requested next review dates are returned in a Map<bookingId, nextReviewDate>