INC-817: Added method to update a single offender next review date
This method gets the needed information from Offender Search API and then
call the `updateMany()` method with the single offender to update.

NOTE: The method makes the request to Offender Search API for simplicity
sake, it would be annoying for the client code to have to make this request
(request handlers/code that handle domain event is already complex enough
as it is).