INC-817: Store next review date in the database (#263)
* INC-817: next_review_date DB table/jpa class and repository

Added a table to store the `nextReviewDate` by `bookingId`.

`bookingId` is the natural primary key (PK) and, unsurprisingly, the other
main columnis the `nextReviewDate` date.

For each `bookingId` there would be only record max, this is a constraint
enforced by the DB because of the fact this is the PK.

We'd likely retrieve "next review dates" for a list of bookingIds.
This is provided by `CoroutineCrudRepository`'s `findAllById()` so we don't
have to add this to the repository when that time comes (I've added a test
for this even tho it's probably overkill as I'd imagine the API for
`CoroutineCrudRepository` is very stable).

* INC-817: Way to get all bookingIds from the reviews table

This is useful in the migration to populate the `next_review_date` table

* INC-817: Class to update next review dates in the database

This will be used to recalculate/update next review dates in the database.

This is necessary when:
1. trying to get a next review date for some bookingIds but some of them
   don't have this value yet (so we need to calculate it)
2. something changes and next review date needs to be updated, e.g.
   a new review is added or someone ACCT status changed

This method gets a list of Offender Search prisoners as this information
is needed to calculate the next review date.

NOTE: At the moment only an `updateMany()` method is available but we can
add another method to `update()` a single offender later on which is
probably what we need in most cases on admission/transfer or when a new
review is created.

* INC-817: Class to get next review dates

This class tries to get these from the database first, for bookingIds
without a `NextReviewDate` DB record it calculate/persist the next review
dates.

After a call to this class all the offenders provided will have a
`NextReviewDate` DB record.

All the requested next review dates are returned in a Map<bookingId, nextReviewDate>

* INC-817: reviews endpoints gets next review dates from the database

The `GET /incentives-reviews/prison/{prisonId}/location/{cellLocationPrefix}`
endpoint now uses the `NextReviewDateGetter` service to get the next
review dates for the list of offenders.

The service will take care of:
1. try to get data from the `next_review_date` database table if possible
2. calculate/update the next review dates for offenders without a date
3. return all the next review dates

NOTE: Point 3 means this endpoint no longer throw a `404 Not Found`
exception as all next review dates for the given offenders are obtained.

* INC-817: Use next review date getter for IepSummary...

When using data from our database. Branches returning data from
NOMIS/Prison API will still calculate it on the fly as before.

NOTE: I've renamed `setNextReviewDate()` to `calculateNextReviewDateFromNomisData()`
to reflect this method is only to be used when data is coming from
NOMIS/Prison API. Once we remove those branches we can remove this and
all code will use the `NextReviewDateGetter` class to get/update next
review dates.

* INC-817: Fix tests by having consistend stub data

Some of the test data was not consistent with each other causing some of
the logic to fail because of unexpected bookingIds

* INC-817: Added method to update a single offender next review date

This method gets the needed information from Offender Search API and then
call the `updateMany()` method with the single offender to update.

NOTE: The method makes the request to Offender Search API for simplicity
sake, it would be annoying for the client code to have to make this request
(request handlers/code that handle domain event is already complex enough
as it is).

* INC-817: Update next review date on sync POST requests

They create a new review which could affect the next review date, so
this should be recalculated.

* INC-817: Update next review date when a new review is added

* INC-817: Update next review date on sync PATCH requests

The iepTime could potentially change and affect the order of the prisoner
reviews, so it makes sense to update the next review date in case this
changed too.

* INC-817: Update next review date on DELETE sync requests

This is a bit of an edge case as reviews wouldn't be deleted often but
when a review is deleted it could potentially affect the next review date.

* INC-817: Update next review date on merge domain events

On merge reviews are copied and persisted with a potentially different
prisonerNumber/bookingId. New `PrisonerIepLevel` records (aka "reviews")
are created, this could have an impact on the next review date and therefore
the next review date is updated.

NOTE: Code was looping through several reviews and saving them one by one.
I've changed the code to do a single `saveAll()` of these records which
should be more efficient as it translate to a single SQL query.

* INC-817: Fix for admission tests missing requests stubs

On admission next review date is updated so these requests need to be
stubbed too.

* INC-817: Tests for NextReviewDateGetterService

* INC-817: Moved `translate()` etc functions in their own file

These are used in different places inside the `service` package to convert
between `PrisonerIepLevel`/`IepDetail`, it makes sense for them to have
their own file and reuse them.

* INC-817: Tests for NextReviewDateUpdaterService

* INC-817: Removed unused `PrisonerIepLevelRepository#getAllBookingIds()`

I thought I'd use it in a migration but it wasn't necessary as we're
updating the next review dates in a lazy way now.

* INC-817: Mark `next_review_date.when_created` as readonly

This field should be automatically set to "now" by postgres and never
changed.

This should avoid the error complaining about "whenCreated" being non-nullable
when trying to update an existing record.

* INC-817: Call `collect()` on `saveAll()`

`saveAll()` returns a `Flow<T>`. Unfortunately this means that the query(ies)
will only be start to be executed when we start to collect.

Annoyingly without collect this does nothing which means it fails silently.

* INC-817: NextReviewDateUpdaterService handle new/existing records

`save()`/`saveAll()` doesn't seem to handle existing records correctly
at the moment unless we explicitly pass the `new =` property to suggest
whether a record should be `INSERT`ed or `UPDATE`d.

NOTE: There may be some clever annotation that could make this automatically
but I haven't found it yet and this should be equivalent (aside from
the slightly more noisy code).

* INC-817: Reverted use of `saveAll()` during MERGE events

It shouldn't make much of a difference but tests currently fail because
the repository call to `saveAll()` would have to be stubbed to return
a non-null `Flow`. It's simpler to leave this method as it was.

* INC-817: PR comment amendment: associate the Kotlin way

Use `k to v` instead of collecting Pair(k, v)

* INC-817: PR amendement: removed unnecessary `var`

* INC-817: PR amendment: use `Flow` returned by `saveAll()`

* INC-817: PR amendement: Removed unnecessary `this` in adapters

* INC-817: PR amendment: Better names for adapters methods

- `translate()` => `toIepDetail()`
- `toMap()` => `toMapByBookingId()`