INC-901: Update next review date on alerts changes
Process domain events of type `prisoner-offender-search.prisoner.updated`
which contains a `categoriesChanged` array which can be used to determine
whether any of the alerts may have changed.

Trigger an update of the next review date when this happen.

NOTE: In the near future there may be a more specific domain event
(e.g. `prisoner-offender-search.prisoner.alerts-updated`) with more information
e.g. `alertCode` (we're only interested about ACCT alerts, which have
`alertCode = "HA"`). So we may revisit/tweak this, but at the moment
this is the closed domain event published.