INC-901: Fixed review domain event description
It was always "An IEP review has been added" even for when a review was
updated or deleted.

NOTE: The SNS service method now takes the description as a parameter which
means this can be used to also publish non-review domain events. This is
also why I've renamed the `reviewId` param, because `AdditionalInformation`
seems to be agnostic about what the ID is for (so it coud be the ID
of something else, e.g. a bookingId)