INC-930: Fix sync DELETE endpoint always assuming review was current
The use of `.let {}` instead of a simple `if ()` caused the the sync DELETE
endpoint to always try to set another review as current even when unnecessary.

This can lead to a violation of the DB constrain that prevents more than
one current reviews for the same `bookindId`.

See: https://mojdt.slack.com/archives/C03BCRW4B5G/p1668786899831079