INC-901: Add integration tests for processing of prisoner.updated messages
The test checks that when a domain event of type `prisoner-offender-search.prisoner.updated`
is received is processed.

The test simulates the scenario when someone was not suitable to be moved
to Standard (confirmed on Basic) but they now have an open ACCT alert,
so the next review date would be in 14 days.