INC-901: Update next review date on alerts changes (#272)
* INC-901: Update next review date on alerts changes

Process domain events of type `prisoner-offender-search.prisoner.updated`
which contains a `categoriesChanged` array which can be used to determine
whether any of the alerts may have changed.

Trigger an update of the next review date when this happen.

NOTE: In the near future there may be a more specific domain event
(e.g. `prisoner-offender-search.prisoner.alerts-updated`) with more information
e.g. `alertCode` (we're only interested about ACCT alerts, which have
`alertCode = "HA"`). So we may revisit/tweak this, but at the moment
this is the closed domain event published.

* INC-901: Fixed review domain event description

It was always "An IEP review has been added" even for when a review was
updated or deleted.

NOTE: The SNS service method now takes the description as a parameter which
means this can be used to also publish non-review domain events. This is
also why I've renamed the `reviewId` param, because `AdditionalInformation`
seems to be agnostic about what the ID is for (so it coud be the ID
of something else, e.g. a bookingId)

* INC-901: Publish domain event when next review date is updated

So that Elastic Search can reindex and potentially other services interested
can subscribe to it too.

* INC-901: PR amendment: Tweaks to log messages

* INC-901: Add integration tests for processing of prisoner.updated messages

The test checks that when a domain event of type `prisoner-offender-search.prisoner.updated`
is received is processed.

The test simulates the scenario when someone was not suitable to be moved
to Standard (confirmed on Basic) but they now have an open ACCT alert,
so the next review date would be in 14 days.

* Revert "INC-901: Publish domain event when next review date is updated"

The current implementation has some issues, we may also understand the
requirement for this better and ultimately this is not part of this
ticket which should focus solely on the updating the next review date on
alerts changes.

This reverts commit a92f9fda0bf60286951edfc0121ecc7df509bd51.